fix after proto update

+++ internal/backend/bssci_v1/structs/messages/msg_vm_ul_data.go
@@ -19,7 +19,7 @@ type VmUlData struct {
 	// ID of the operation
 	OpId int64 `msg:"opId" json:"opId"`
 	// MAC-Type of the Variable MAC
-	MacType int64 `msg:"macType" json:"macType"`
+	MacType uint8 `msg:"macType" json:"macType"`
 	// n Byte End Point user data U-MPDU; starting with first byte after MAC-Type
 	UserData []uint8 `msg:"userData" json:"userData"`
 	// Transceiver time of reception, center of last subpacket, 64 bit, ns resolution
@@ -48,7 +48,7 @@ type VmUlData struct {
 
 func NewVmUlData(
 	opId int64,
-	macType int64,
+	macType uint8,
 	userData []byte,
 	trxTime uint64,
 	freqOff float64,
@@ -116,7 +116,7 @@ func (m *VmUlData) IntoProto(bsEui *common.EUI64) *bs.EndnodeUplink {
 		Message: &bs.EndnodeUplink_VmUlData{
 			VmUlData: &bs.EndnodeVariableMacUlDataMessage{
 				Data:      m.UserData,
-				MacType:   m.MacType,
+				MacType:   uint32(m.MacType),
 				FreqOff:   m.FreqOff,
 				CarrSpace: bs.CarrierSpacingEnum(m.CarrSpace),
 				PattGrp:   bs.TsmaPatternGroupEnum(m.PattGrp),
